% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_TransiTypha.R
\name{barplot_apparitions}
\alias{barplot_apparitions}
\title{barplot_apparitions - trace pour l'ensemble des stations apparues, les proportions par type}
\usage{
barplot_apparitions(
  APPAR,
  MATTRANS,
  POS = "left",
  XLIM = c(0, 100),
  XLAB = "Proportion des apparitions (\%)",
  MAIN = NULL,
  COUL = NULL,
  LISTCOUL = c("#FFCCB6", "#B9B9B9", "#FC8D1D", "#2DB30C", "#305B25", "#EAC931",
    "#977D09", "#606060")
)
}
\arguments{
\item{APPAR}{le nombre d'apparition par type, calculé avec mattrans.calc cf $Nb_App}

\item{MATTRANS}{la matrice de transition calculée avec mattrans.calc cf $Matrice_transition}

\item{POS}{"left" ou "right" donne la position des noms et nombre des types..}

\item{XLIM}{permet de changer les limites de l'axe des x doit être sous format c(n,n)}

\item{XLAB}{permet de changer le titre de l'axe des x doit être entre guillemets}

\item{MAIN}{permet de changer le titre du graphique doit être entre guillemets}

\item{COUL}{permet de changer les couleurs des barplot}

\item{LISTCOUL}{permet de donner la liste complète des couleurs de tous les types (y compris ceux absents du graphe..)}
}
\value{

}
\description{
barplot_apparitions - trace pour l'ensemble des stations apparues, les proportions par type
}
\examples{
# Exemple avec jeu de données inclu dans le package
data_ty

# contient colonnes avec les ID des stations, les Dates, les longueurs, les limons,
# l'abondance arbustif+ arboré, la fréquence et 4 colonnes pour les données de suivi..

# On spécifie la partie ‘suivi’ (ici les colonnes 7 à 10) :
suivi <- data_ty[,7:10]

Typhamat_all <- mattrans.calc(TAILLE = data_ty$Longueur, FREQ = data_ty$Freq,
                               ARB = data_ty$ArbuArbo, LIMONS = data_ty$Limon,
                               ANNEES = data_ty$Dates, IDSTATION = data_ty$ID, SUIVI = suivi)
# On peut voir la Classification des stations
Typhamat_all$Classification
data.frame(station = data_ty$ID, Dates = data_ty$Dates,
           Classification = Typhamat_all$Classification)
# On peut voir la matrice de transition
Typhamat_all$Matrice_transition

# Pour tracer le bilan des apparitions
barplot_apparitions(APPAR = Typhamat_all$Nb_App, MATTRANS = Typhamat_all$Matrice_transition)

}
