% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_TransiTypha.R
\name{Typha_transition_graph}
\alias{Typha_transition_graph}
\title{Typha_transition_graph - permet de tracer les états et les transitions}
\usage{
Typha_transition_graph(
  MATTRANS,
  CLASSIF,
  ANNEES,
  LISTCOUL = c("#FFCCB6", "#B9B9B9", "#FC8D1D", "#2DB30C", "#305B25", "#EAC931",
    "#977D09", "#606060"),
  WIDTH = 15,
  DEL = 0.1,
  CURVE = 0.2,
  ARROW = 0.7,
  SIZE = 100,
  LABELS = NULL
)
}
\arguments{
\item{MATTRANS}{la matrice de transition calculée avec mattrans.calc cf $Matrice_transition}

\item{CLASSIF}{la classification calculée avec mattrans.calc cf $Classification}

\item{ANNEES}{vecteur de même dimension que IDSTATION, l'année de visite de la station}

\item{LISTCOUL}{(facultatif) liste de couleurs à utiliser pour les types de stations}

\item{WIDTH}{(facultatif) épaisseur des transitions}

\item{DEL}{(facultatif) la proportion minimale de transition pour faire apparaître un lien (doit être compris entre 0 et 1)}

\item{CURVE}{(facultatif) courbure des transitions}

\item{ARROW}{(facultatif) taille des flèches}

\item{SIZE}{(facultatif) taille des noeuds (qui sera de toutes façons proportionnels au nombre de stations de départ par type)}

\item{LABELS}{(facultatif) étiquettes des noeuds}
}
\value{

}
\description{
Typha_transition_graph - permet de tracer les états et les transitions
}
\examples{
# Exemple avec jeu de données inclu dans le package
data_ty

# contient colonnes avec les ID des stations, les Dates, les longueurs, les limons,
# l'abondance arbustif+ arboré, la fréquence et 4 colonnes pour les données de suivi..

# On spécifie la partie ‘suivi’ (ici les colonnes 7 à 10) :
suivi <- data_ty[,7:10]

Typhamat_all <- mattrans.calc(TAILLE = data_ty$Longueur, FREQ = data_ty$Freq,
                               ARB = data_ty$ArbuArbo, LIMONS = data_ty$Limon,
                               ANNEES = data_ty$Dates, IDSTATION = data_ty$ID, SUIVI = suivi)
# On peut voir la Classification des stations
Typhamat_all$Classification
data.frame(station = data_ty$ID, Dates = data_ty$Dates,
           Classification = Typhamat_all$Classification)
# On peut voir la matrice de transition
Typhamat_all$Matrice_transition

# On peut tracer le graphe des transitions :
Typha_transition_graph(MATTRANS = Typhamat_all$Matrice_transition,
                       CLASSIF = Typhamat_all$Classification,
                       ANNEES = data_ty$Dates, SIZE = 150)
}
